+++ tests/som/tests/BasicInterpreterTests.java
@@ -84,8 +84,8 @@ public class BasicInterpreterTests {
 
   protected void assertEqualsSOMValue(final Object expectedResult, final Object actualResult) {
     if (resultType == SInteger.class) {
-      int expected = ((java.lang.Integer) expectedResult).intValue();
-      int actual   = ((SInteger) actualResult).getEmbeddedInteger();
+      long expected = (Long) expectedResult;
+      long actual   = ((SInteger) actualResult).getEmbeddedInteger();
       assertEquals(expected, actual);
       return;
     }
